divide ssr file
